+++ packages-fetcher/fetcher-helper-file/src/helper/prepare-form-data.ts
@@ -4,6 +4,9 @@ import {
 
 import makeSureFile from './make-sure-file';
 
+/**
+ * 创建一个 FormData，并填入以 formFieldName 为名（默认值为 file）的 File 对象
+ */
 export default async function prepareFormData(file: TFileAlike, formFieldName = 'file'): Promise<FormData> {
   const formData = new FormData();