+++ packages-ui/file-picker/CHANGELOG.md
@@ -1,12 +1,16 @@
 # CHANGELOG
 
+## 0.1.2 2026/06/12 @驳是
+
+* FIX `display: none !important` to avoid stupid CSS from other UI libs
+
 ## 0.1.1 2026/06/04 @驳是
 
 * FEAT Add `props.disabled`
 
 ## 0.1.0 2026/06/03 @驳是
 
-* FIX 修复 Chrome 和 Safari 下选择后没有回调的问题
+* FIX Chrome & Safari will NOT fire change due to they mute original value
 
 ## 0.0.1 2026/06/01 @驳是